reactor: separate favorite dukan logic form dukan service to FavouriteDukanService
